MCQ BUILDER/requirements.txt
langchain==0.1.4
google-generativeai==0.3.2
streamlit==1.30.0
langchain-google-genai
python-dotenv==1.0.1
python-docx==0.8.11
google-auth==2.29.0
firebase-admin==6.5.0
pypdf
